+++ lib/pages/home_Widgets/catalog_item.dart
@@ -1,13 +1,11 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 //CatalogItem
 import 'package:flutter/material.dart';
+import 'package:thirty_days_of_flutter/pages/home_Widgets/add_to_cart.dart';
 import 'package:velocity_x/velocity_x.dart';
-
 import 'package:thirty_days_of_flutter/main.dart';
-import 'package:thirty_days_of_flutter/models/cart.dart';
 import 'package:thirty_days_of_flutter/models/catalog.dart';
 import 'package:thirty_days_of_flutter/pages/home_Widgets/catalog_image.dart';
-import 'package:thirty_days_of_flutter/pages/home_page.dart';
 
 class CatalogItem extends StatelessWidget {
   final Item catalog;
@@ -39,44 +37,12 @@ class CatalogItem extends StatelessWidget {
               buttonPadding: EdgeInsets.zero,
               children: [
                 "\$${catalog.price}".text.bold.xl.make(),
-                _AddToCart(catalog: catalog)
+                AddToCart(catalog: catalog)
               ],
             ).pOnly(right: 8.0)
           ],
         ))
       ],
     )).white.rounded.square(150).make().py16();
-  }
-}
-
-class _AddToCart extends StatefulWidget {
-  final Item catalog;
-  const _AddToCart({
-    Key? key,
-    required this.catalog,
-  }) : super(key: key);
-  @override
-  State<_AddToCart> createState() => _AddToCartState();
-}
-
-class _AddToCartState extends State<_AddToCart> {
-  bool isAdded = false;
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-        onPressed: () {
-          isAdded = isAdded.toggle();
-          final _catalog = CatalogModel();
-          final _cart = CartModel();
-          _cart.catalog = _catalog;
-          _cart.add(widget.catalog);
-          setState(() {});
-        },
-        style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all(MyApp.darkBluesh),
-            shape: MaterialStateProperty.all(const StadiumBorder())),
-        child: isAdded
-            ? const Icon(Icons.done)
-            : "Add to Cart".text.color(Colors.white).make());
   }
 }
